# pp-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.AddRoomMemberRequest:
    properties:
      member_username:
        type: string
    required:
    - member_username
    type: object
...
  handler.ChatRoomResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.CreateChatRoomRequest:
    properties:
      description:
        type: string
      name:
        type: string
      type:
        enum:
        - public
        - private
        - group
        type: string
    required:
    - name
    - type
    type: object
...
  handler.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
  handler.RoomMemberResponse:
    properties:
      joined_at:
        type: string
      member_username:
        type: string
      room_id:
        type: string
    type: object
...
  handler.UpdateChatRoomRequest:
    properties:
      description:
        type: string
      name:
        type: string
      type:
        enum:
        - public
        - private
        - group
        type: string
    required:
    - name
    - type
    type: object
...
  handler.UserResponse:
    properties:
      created_at:
        type: string
      is_active:
        type: boolean
      last_online_at:
        type: string
      nickname:
        type: string
      profile_picture_url:
        type: string
      role:
        type: string
      status_message:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pitturu.com
    name: PittuRu Team
    url: https://pitturu.com
  description: PittuRu 게임 플랫폼 REST API 서버
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PittuRu API
  version: "1.0"
paths:
...
  /chat-rooms:
    get:
      description: Retrieves a list of all chat rooms.
      parameters:
      - description: Limit the number of rooms returned
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.ChatRoomResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: List all chat rooms
      tags:
      - Chat Rooms
    post:
      consumes:
      - application/json
      description: Creates a new chat room.
      parameters:
      - description: Chat room details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/handler.CreateChatRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.ChatRoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Create a new chat room
      tags:
      - Chat Rooms
  /chat-rooms/{room_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a chat room by its ID. Only members or admin can delete.
      parameters:
      - description: Chat Room ID
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Delete a chat room
      tags:
      - Chat Rooms
    get:
      description: Retrieves a single chat room by its unique ID.
      parameters:
      - description: Chat Room ID
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ChatRoomResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: Get chat room by ID
      tags:
      - Chat Rooms
    put:
      consumes:
      - application/json
      description: Updates an existing chat room by its ID. Only members or admin
        can update.
      parameters:
      - description: Chat Room ID
        in: path
        name: room_id
        required: true
        type: string
      - description: Updated chat room details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateChatRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ChatRoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Update a chat room
      tags:
      - Chat Rooms
  /chat-rooms/{room_id}/members:
    get:
      description: Retrieves a list of members for a specific chat room.
      parameters:
      - description: Chat Room ID
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      summary: List room members
      tags:
      - Chat Rooms
    post:
      consumes:
      - application/json
      description: Adds a user as a member to a chat room.
      parameters:
      - description: Chat Room ID
        in: path
        name: room_id
        required: true
        type: string
      - description: Member details
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/handler.AddRoomMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RoomMemberResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Add room member
      tags:
      - Chat Rooms
  /chat-rooms/{room_id}/members/{username}:
    delete:
      consumes:
      - application/json
      description: Removes a user from a chat room. Only the member themselves or
        admin can remove.
      parameters:
      - description: Chat Room ID
        in: path
        name: room_id
        required: true
        type: string
      - description: Member username to remove
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Remove room member
      tags:
      - Chat Rooms
...
  /me/chat-rooms:
    get:
      description: Retrieves a list of chat rooms the authenticated user is a member
        of.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.ChatRoomResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: List user chat rooms
      tags:
      - Chat Rooms
...
  /ws/chat:
    get:
      description: Upgrades the HTTP connection to a WebSocket for real-time chat.
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Establish WebSocket connection for chat
      tags:
      - Chat
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"